Avoid excluding LVM slaves when getting valid block devices (bsc#1194299)
Function getValidBlockDevices() implements special handling for
device-mapper devices. It primarily tries to recognize multipath DMs
which then treats as real disks and skips their slaves. This is because
multipath is a request-based DM target and so block attributes managed
by saptune (such as the elevator) should be applied to the parent device
to take effect. Other DM devices and their slaves are then
conservatively all excluded by the function.

The patch updates getValidBlockDevices() to skip DM slaves only for
multipath devices. Other DM types can have their slaves treated normally
because they stack on top of "normal" block devices and do not take over
their I/O scheduling.

This makes sure that if an LVM device is built on top of a whole disk
(e.g. sda) then getValidBlockDevices() returns the disk in its result
and subsequently saptune will tune its attributes according to the
requested notes.

Notice that if an LVM device was previously built on top of a partition
(sda1) that covered a whole disk (sda) then getValidBlockDevices()
already did not exclude this disk. The patch makes handling of these two
cases (LVM on top of a whole disk vs. on top of a partition) consistent.